Falsh and RAM saving, Add GPS speed

+++ trunk/src/pers.cpp
@@ -42,14 +42,14 @@ void generalDefault()
   g_eeGeneral.contrast = 25;
   g_eeGeneral.vBatWarn = 90;
   g_eeGeneral.stickMode=  1;
-  for (int i = 0; i < 7; ++i) {
+  for (uint8_t i = 0; i < 7; ++i) {
     g_eeGeneral.calibMid[i]     = 0x200;
     g_eeGeneral.calibSpanNeg[i] = 0x180;
     g_eeGeneral.calibSpanPos[i] = 0x180;
   }
   strncpy_P(g_eeGeneral.ownerName,PSTR("ME        "),10);
   int16_t sum=0;
-  for(int i=0; i<12;i++) sum+=g_eeGeneral.calibMid[i];
+  for(uint8_t i=0; i<12;i++) sum+=g_eeGeneral.calibMid[i];
   g_eeGeneral.chkSum = sum;
 }
